refactor: better structure
